Add make_model_metadata_of_size to testing_utils.
